Remove [Internal] [Database Initialization] [Spinner] "tea.KeyMsg" (#318)
- [+] refactor(setup.go): remove unnecessary tea.KeyMsg case in Update method

Note: This should work fine and be safe. It won't cause any hangs or deadlocks, even when waiting for a response from the server while connecting (on my way) to the database in this goroutine. Most databases already have a default timeout set in their configuration. If wanting to explicitly set a timeout, it must be set in the server configuration before this goroutine starts connecting to the database.