Docs [Crypto] [Hybrid Scheme] Update Docs Related nonce capacity (#393)
* Fix Typo

- [+] chore(authentication): rename chiper.go to cipher.go

* Docs [Crypto] [Hybrid Scheme] Update Docs Related nonce capacity

- [+] feat(stream): add constant for additional nonce capacity percentage

Note: The additionalCapacityPercentage constant is introduced to represent the percentage of additional capacity to be added to the nonce capacity when it exceeds the minimum required size. This constant is used in both the AESNonceCapacity and ChachaNonceCapacity methods to calculate the nonce capacity for AES-CTR and XChaCha20-Poly1305 respectively. The default value is set to 0.05, which means an additional 5% capacity will be added.