Implement [Crypto] [Hybrid Scheme] Customize Nonce Capacity (#395)
- [+] feat(stream): add CustomizeNonceCapacity method to customize nonce capacity

Note: The CustomizeNonceCapacity method allows customizing the nonce capacity for AES-CTR and XChaCha20-Poly1305 in the Stream struct. It introduces a new CustomizeCapacityNonce struct to store the customized nonce capacities. The default value for both AESNonceCapacity and ChachaNonceCapacity is set to 0.05 (5% additional capacity). The calculateAESNonceCapacity and calculateChachaNonceCapacity methods are updated to use the customized nonce capacities. A new test case TestEncryptWithCustomizeNonceCapacity is added to verify the functionality of the CustomizeNonceCapacity method.