Docs [Crypto] Update Documentation (#447)
- [+] feat(stream): add package documentation for hybrid encryption stream

The package documentation provides an overview of the stream package, including its purpose, compatibility considerations, and recommended usage scenarios for secure encryption and decryption of data streams using a hybrid approach combining AES-CTR and XChaCha20-Poly1305 algorithms.

- [+] test(boringtls): add compatibility note for stream encryption in test

In the TestStreamServerWithCustomTransport test, a note is added to clarify that while the automatic encryption/decryption during transport works well for internal Go applications, it may not be compatible with front-end apps such as browsers due to the specific cipher used. The note also mentions that even with TLS 1.3, not all browsers will work properly if used for HTTPS front-end, with Firefox being able to encrypt but unable to decrypt.