Enchance [Key Auth + Session Middleware Logic] (#720)
* Enchance [Key Auth + Session Middleware Logic]

- [+] feat(authentication): add Signature field to AuthorizationData struct

Note: The commit message indicates that a new field called `Signature` has been added to the `AuthorizationData` struct. This field is of type `any` and is used to store signature data associated with the API key. The signature data can be used for enhanced security purposes or other purposes. The commit also updates the documentation to reflect the addition of the `Signature` field and removes the TODO comment regarding handling ECDSA signatures.

* Docs [Key Auth + Session Middleware Logic] Update Comments

- [+] refactor(constant.go): update comment to mention Valkey as an alternative to Redis for faster response times