Implement Skip Middleware & Helper Function "CustomNextPathAvailable"  (#752)
* Feat [Middleware Utils] "NewSkipMiddleware"

- [+] feat(middleware): add NewSkipMiddleware function

Note: The NewSkipMiddleware function creates a new middleware handler that skips the wrapped handler based on the provided exclude predicate function. If the exclude predicate returns true, the wrapped handler will be skipped, and the request will be passed to the next middleware in the chain using c.Next. If the exclude predicate returns false, the wrapped handler will be executed.

* Feat [Middleware] [Helper Function] "CustomNextPathAvailable"

- [+] feat(middleware): add CustomNextPathAvailable helper function

Note: The CustomNextPathAvailable function creates a custom Next function for Fiber middleware. It checks if the requested path is available in the provided map of paths. If the path is found, the Next function returns false, indicating that the middleware should not be skipped. Otherwise, it returns true, indicating that the middleware should be skipped.