Refactor [Frontend] [HTMX+TEMPL] [MIME] Render Content (#783)
- [+] refactor(error_page_handler): set Content-Type header based on presence of non-ASCII characters

Note: The Content-Type header is now set dynamically based on the presence of non-ASCII characters in the rendered HTML response. If non-ASCII characters are present, the MIME type with charset (fiber.MIMETextHTMLCharsetUTF8) is used. Otherwise, if the response body contains only ASCII characters, the MIME type without charset (fiber.MIMETextHTML) is used.