Enhance [REST APIs Router] Redirect Endpoint
- [+] feat(restapis_routes): add redirect rules for health check endpoints

Note: This commit enhances the redirect middleware by adding redirect rules for health check endpoints. If a request is made to `/v1/server/health` or `/v1/server/health/`, it will be redirected to the root `/` endpoint. This is a technique to manipulate bot scanners attempting to access sensitive directories.
Additionally, the redirect middleware is now applied to all HTTP methods (GET, POST, PUT, DELETE, PATCH) for the `*` path pattern, ensuring comprehensive coverage of potential scanner requests.