Implement [Middleware] [Helper Function] "CustomNextStatusCode" (#790)
- [+] feat(middleware): add CustomNextStatusCode helper function

Note: The CustomNextStatusCode helper function creates a custom Next function for the Fiber middleware. It checks the HTTP status code of the response and determines whether to skip the middleware based on the provided status codes. If the response status code matches any of the specified status codes, the Next function returns true, indicating that the middleware should be skipped.

- [+] feat(routes): add status code skipper for cache middleware
- [+] Add a status code skipper using the CustomNextStatusCode helper function to skip the cache middleware for specific status codes. In this case, the cache middleware will be skipped for the 301 (Moved Permanently) status code.
- [+] refactor(routes): use CustomNextStack for cache middleware skipper
- [+] Refactor the cache middleware skipper to use the CustomNextStack function, which allows combining multiple skipper functions. The contentTypeSkip and statusCodeSkip functions are passed as a map to the CustomNextStack function to create a combined skipper for the cache middleware.